downloader: Move cheap Mercurial functional tests to unit tests
